Remove now obsolete versions of RainbowFit
